+++ upkie/cpp/actuation/static_config.h
@@ -57,15 +57,17 @@ inline moteus::QueryCommand query_resolution() noexcept {
  * \return Upkie's servo layout.
  */
 inline const ServoLayout servo_layout() noexcept {
-  constexpr int kLeftBusID = 1;   // JC1 on the pi3hat
-  constexpr int kRightBusID = 3;  // JC3 on the pi3hat
+  constexpr int kLeftBusID = 1;              // JC1 on the pi3hat
+  constexpr int kRightBusID = 3;             // JC3 on the pi3hat
+  constexpr double kMaxQdd100Torque = 16.0;  // [N m]
+  constexpr double kMaxMj5208Torque = 1.7;   // [N m]
   ServoLayout layout;
-  layout.add_servo(1, kLeftBusID, "left_hip");
-  layout.add_servo(2, kLeftBusID, "left_knee");
-  layout.add_servo(3, kLeftBusID, "left_wheel");
-  layout.add_servo(4, kRightBusID, "right_hip");
-  layout.add_servo(5, kRightBusID, "right_knee");
-  layout.add_servo(6, kRightBusID, "right_wheel");
+  layout.add_servo(1, kLeftBusID, "left_hip", kMaxQdd100Torque);
+  layout.add_servo(2, kLeftBusID, "left_knee", kMaxQdd100Torque);
+  layout.add_servo(3, kLeftBusID, "left_wheel", kMaxMj5208Torque);
+  layout.add_servo(4, kRightBusID, "right_hip", kMaxQdd100Torque);
+  layout.add_servo(5, kRightBusID, "right_knee", kMaxQdd100Torque);
+  layout.add_servo(6, kRightBusID, "right_wheel", kMaxMj5208Torque);
   return layout;
 }